requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
requests==2.31.0
Pillow==10.0.0
numpy==1.24.3
